显示ComputeType

+++ jobserver-common/src/main/java/io/github/melin/flink/jobserver/core/entity/FlinkDriver.java
@@ -62,7 +62,7 @@ public class FlinkDriver implements IEntity {
     private Integer instanceCount;
 
     @Column(name = "server_cores", nullable = false)
-    private Long serverCores;
+    private Integer serverCores;
 
     @Column(name = "server_memory", nullable = false)
     private Long serverMemory;
@@ -113,7 +113,7 @@ public class FlinkDriver implements IEntity {
         jobServer.setGmtCreated(Instant.now());
         jobServer.setGmtModified(Instant.now());
         jobServer.setInstanceCount(0);
-        jobServer.setServerCores(0L);
+        jobServer.setServerCores(0);
         jobServer.setServerMemory(0L);
         jobServer.setShareDriver(shareDriver);
         jobServer.setLogServer(hostName);
